requirements.txt
click==8.1.8
crewai>=0.130.0
crewai-tools>=0.47.1
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pypdf==4.2.0
openai>=1.68.2
langchain-openai>=0.1.0
pandas==2.2.2
pillow==10.3.0
